Move static DB seeds into :app:db:seed (run during Schema.create)
Create :app:db:seed (sqldelight, dependency(:schema), pkg com.moneymanager.database.sql.seed)
holding seed INSERTs that run during Schema.create AFTER the schema module's tables and
triggers. The checked-in StaticSeed.sq seeds the lookup tables (audit_type, source_type,
api_session_type, platform), the well-known attribute/relationship types, the SYSTEM device
at a fixed id (WellKnownIds.SYSTEM_DEVICE_ID = 1), and the Uncategorized category with its
SYSTEM provenance. Audited rows (the category) still fire their audit triggers, exactly as
the old runtime seeding did.

The Jvm/Android DatabaseManagers now create the seed module's aggregated Schema; the static
block is removed from DatabaseConfig.seedDatabase (currencies + built-in strategies still
seed in Kotlin for now — moved to build-time-generated .sq in a follow-up). db:core test
suite passes.
